requirements-experiments.txt
cma==4.4.4
numpy==2.3.5
scipy==1.16.3
pandas==2.3.3
matplotlib==3.10.8
statsmodels==0.14.6
PyYAML==6.0.2
